+++ contracts/hackatom/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod contract;
 
-/** Below we expose wasm exports **/
 #[cfg(target_arch = "wasm32")]
-pub use cosmwasm_std::{allocate, deallocate};
+#[allow(unused_imports)]
+use cosmwasm_std; // Pull in cosmwasm_std's C externs (cosmwasm_vm_version_1, allocate, deallocate)
 
 #[cfg(target_arch = "wasm32")]
 mod wasm {
@@ -11,7 +11,7 @@ mod wasm {
     use std::ffi::c_void;
 
     #[no_mangle]
-    pub extern "C" fn init(env_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
+    extern "C" fn init(env_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
         do_init(
             &contract::init::<ExternalStorage, ExternalApi>,
             env_ptr,
@@ -20,7 +20,7 @@ mod wasm {
     }
 
     #[no_mangle]
-    pub extern "C" fn handle(env_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
+    extern "C" fn handle(env_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
         do_handle(
             &contract::handle::<ExternalStorage, ExternalApi>,
             env_ptr,
@@ -29,7 +29,7 @@ mod wasm {
     }
 
     #[no_mangle]
-    pub extern "C" fn query(msg_ptr: *mut c_void) -> *mut c_void {
+    extern "C" fn query(msg_ptr: *mut c_void) -> *mut c_void {
         do_query(&contract::query::<ExternalStorage, ExternalApi>, msg_ptr)
     }
 }